Mistakes returned to int

+++ Homework 1/src/main/java/ru/ifmo/ctddev/ml/homework1/KNNAlgorithm.java
@@ -34,12 +34,12 @@ public class KNNAlgorithm {
         int optimalK = -1;
         double minAverageMistakes = -1.0;
         for (int i = 0; i < FOLD_NUMBER; i++) {
-            double sumMistakes = 0;
+            int sumMistakes = 0;
             for (int j = 0; j < i; j++) {
                 splitter.split(i);
                 sumMistakes += countMistakes(splitter.getTrainingDataSet(j), splitter.getTestingDataSet(j), i);
             }
-            double averageMistakes = sumMistakes / (double) i;
+            double averageMistakes = (double) sumMistakes / (double) i;
             if (optimalK == -1 || averageMistakes < minAverageMistakes) {
                 optimalK = i;
                 minAverageMistakes = averageMistakes;
@@ -48,8 +48,8 @@ public class KNNAlgorithm {
         return optimalK;
     }
 
-    private double countMistakes(List<DataSetEntity> trainingDataSet, List<DataSetEntity> testingDataSet, int nearestPointsNumber) {
-        double mistakes = 0;
+    private int countMistakes(List<DataSetEntity> trainingDataSet, List<DataSetEntity> testingDataSet, int nearestPointsNumber) {
+        int mistakes = 0;
         for (DataSetEntity test : testingDataSet) {
             List<DataSetDistance> distances = trainingDataSet.stream()
                     .map(trainingSet -> new DataSetDistance(distanceCounter.apply(trainingSet.getFeature(), test.getFeature()), trainingSet.getEntityClass()))
